Rename motor controller wrappers
Fixes #114
