+++ core/src/main/java/com/chopshop166/chopshoplib/motors/WTalonBase.java
@@ -10,7 +10,7 @@ import edu.wpi.first.wpilibj.motorcontrol.MotorController;
 /**
  * Talon that is compatible with the Smart Motor Controller.
  */
-public class PIDTalonBase<T extends BaseTalon & MotorController & Sendable> extends SmartMotorController {
+public class WTalonBase<T extends BaseTalon & MotorController & Sendable> extends SmartMotorController {
 
     /** Reference to the wrapped Talon. */
     private final T wrapped;
@@ -23,7 +23,7 @@ public class PIDTalonBase<T extends BaseTalon & MotorController & Sendable> exte
      * @param wrapped    The raw object to wrap.
      * @param resolution The number of ticks per revolution.
      */
-    public PIDTalonBase(final T wrapped, final double resolution) {
+    public WTalonBase(final T wrapped, final double resolution) {
         super(wrapped, new TalonEncoder(wrapped, resolution));
         this.wrapped = wrapped;
     }
